collector/requirements.txt
requests
beautifulsoup4
lxml
python-dateutil
# Optionnels mais recommandés
feedparser
pdfminer.six
